requirements.txt
torch==1.9.0
numpy==1.21.0
Pillow==8.3.1
facenet-pytorch==2.5.2
scikit-learn==0.24.2